+++ app.py
@@ -1,19 +1,15 @@
-#!/usr/local/bin/python3.6
 from flask import Flask
 from flask import render_template, redirect, request
-import random as r
-import string
-import os
+from utils import *
 
 app = Flask(__name__)
 
-uploads_dir = os.path.join(os.getcwd(), "uploads")
-length_of_id = 7
-id_chars = string.digits + string.ascii_letters
-
 @app.route("/paste/<paste_id>", methods=['GET'])
 def get_paste(paste_id):
-  return render_template("200.html", paste=)
+  body = read_paste(paste_id)
+  if not body:
+    abort(404)
+  return render_template("200.html", paste=body)
 
 @app.route("/", methods=['POST', 'GET'])
 def post_paste():
@@ -21,32 +17,12 @@ def post_paste():
   if method == "GET":
     return render_template("home.html")
   elif method == "POST":
-    paste_id = store(request.body)
+    paste_id = store(request.form['paste'])
     return redirect("/paste/{}".format(paste_id))
-  else:
-    return render_template("405.html")
 
-def store(paste):
-  paste_id = generate_paste_id()
-  store_path = os.path.join(uploads_dir, paste_id)
-  with open(store_path, "w") as f:
-    f.write(paste)
-  return paste_id
-
-def create_paste_id():
-  return "".join(r.choices(id_chars, k=length_of_id))
-
-def generate_paste_id():
-  paste_id = create_paste_id()
-  store_path = get_file_path(paste_id)
-  while os.path.exists(store_path):
-    paste_id = create_paste_id()
-  return paste_id
-  
-def read_paste()
-    
-app.run(
+if __name__ == "__main__":
+  check_uploads_folder()
+  app.run(
     debug=True,
     host='0.0.0.0',
-    port=10000
-)
+    port=8080)
