+++ templates/200.html
@@ -1,5 +1,8 @@
 <html>
   <body>
-    {% paste %}
+    {{ paste }}
   </body>
+  <br/>
+  <br/>
+  <a href="/">back</a>
 </html>
