+++ templates/home.html
@@ -1,7 +1,8 @@
 <html>
   <body>
-    <form action="/upload" method="POST">
-      <input type="textarea" required>Paste here</input></br>
+    <form action="/" method="POST">
+      <label for="paste">Paste here!</label><br>
+      <input type="textarea" rows=100 cols=100 name="paste" required></input></br>
       <input type="submit"/>
     </form>
   </body>
